* Cloned the Registration code to give a foundation for the Login.
* Merged the Login/Register Controllers into one AccountController.
