--- src/Mentat.UI/Controllers/LoginController.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Mentat.UI.Controllers
-{
-    public class LoginController : Controller
-    {
-
-        [HttpGet]
-        public string Index()
-        {
-            return "test";
-        }
-    }
-}
