+++ src/Mentat.UI/ViewModels/LoginViewModel.cs
@@ -1,0 +1,12 @@
+﻿namespace Mentat.UI.ViewModels;
+
+
+public class LoginViewModel
+{
+    public string UserName { get; set; }
+    
+    public string Password { get; set; }
+
+    public int UserType { get; set; } // 0 for students, 1 for mentors
+
+}
