+++ src/Mentat.UI/Views/Shared/_Layout.cshtml
@@ -38,8 +38,8 @@
                           <img src="~/images/blank-profile-picture-g235ada478_640.png" width="40" height="40" class="rounded-circle">
                         </a>
                         <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                          <a class="dropdown-item" asp-area="" asp-controller="Register" asp-action="Register">Register</a>
-                          <a class="dropdown-item" asp-area="" asp-controller="Login" asp-action="Index">Log In</a>
+                          <a class="dropdown-item" asp-area="" asp-controller="Account" asp-action="Register">Register</a>
+                          <a class="dropdown-item" asp-area="" asp-controller="Account" asp-action="Login">Log In</a>
                         </div>
                       </li>   
                     </ul>
